requirements.txt
streamlit==1.32.0
supabase==2.3.5
crewai==0.16.2
python-dotenv==1.0.1
google-generativeai==0.3.2
Pillow==10.2.0
requests==2.31.0
gdown==5.1.0 